# feat: optimize resize-osdrive replacing get-partitionsupportedsize with diskpart